move isGameReadyToStart to store
